requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
docker==7.0.0

# E2B Cloud Sandboxes
e2b-code-interpreter==1.0.0

# LLM Providers
google-genai>=1.0.0

# Environment Variables
python-dotenv==1.0.0

# Data Handling
pydantic==2.5.3
pyyaml==6.0.1

# Optional: Development Tools
# pytest==7.4.4
# black==24.1.1
# flake8==7.0.0
